added support for non-button touch pads for MM game engine
There are some systems with touch pads (like Mac notebooks) which do not
have separate left/right buttons, which are required for controlling the
Mirror Magic game engine (which usually works best with left/right mouse
buttons).

This change adds support for recognizing touch clicks on the left and
right half of the touch pad as left and right button clicks to rotate
mirrors and other rotatable game elements in the MM game engine to the
left or to the right.