fixed network games, part 1: separate locally and network connected players
Network games were largely broken since (at least) version 3.2.4
(although they still worked for the most common cases, but only for the
first game after starting the program). They will be fixed in several
separate steps.

This first part separates storing if a team-mode player was connected
locally or over the network (and then sets the previously used general
flag if a certain player is "connected" or not, regardless of being
connected locally or remotely).