fixed network games, part 2: set player focus after setting local player
There was also a bug with setting the focussed player to the local
player where the local player potentially changed afterwards (if the
player for network games was assigned to any player in the playfield
different from the player that was selected for network games by the
user on program start).

To fix this, the code to set the focussed player for network games was
moved to a point after setting the local player.