fixed network games, part 3: enabled mapping input to reassigned players
Before this change, player input from connected network players was
only allowed for active players. However, the old code did not take
into account that players may have been reassigned to other players
in the level playfield, so their input actions have to be mapped to
that different player.

This change accepts input actions for all players, and later maps
them to the finally assigned player in the game.