added setup option to enable network multi-player games
This is a work-in-progress change: The new setup option to enable
network mode just sets the same flag as using the command-line option
"--network" (without specifying a hostname or a port).

For local network games, and if UDP broadcasting works as it should
to auto-detect a local network server that is already running, this is
already sufficient to play network multi-player games.

What's still missing is a way to explicitly specify a network server
to also be able to play network games with hosts outside of the local
network (or if UDP broadcasting does not work correctly).