added "passthrough" option for clickable global animations
This change adds a new option for the "graphicsinfo.conf" file for
defining a "passthrough" option for global animations that causes
clicks on clickable global animations not to be consumed, but to
be passed-through to the underlying screen controls (or other global
animations under the clicked global animation). The new option works
like this:

global.anim_1.part_1.MAIN.anim_event:            click
global.anim_1.part_1.MAIN.style:                 passthrough_clicks

The second option is the new one (the first one already existed).
When clicking this global animation (on the main menu screen), the
click will passed-through to the main menu screen to be further
processed by screen controls like gadgets or menu buttons.