fixed order of handling mouse clicks on gadgets and on global animations
Before, gadgets were checked for clicks before global animations, which
caused clickable global animations to not be clicked if a gadget at the
same position was successfully clicked (because mouse events are forced
to be off-screen afterwards).

Now, clickable global animations are checked for mouse clicks before
gadgets, and they either consume the mouse click (so a gadget at the
same position will not receive the mouse click anymore), or they are
set to "passthrough_clicks" style (so a gadget at the same position
will still receive the mouse click).