fixed bug causing crash when trying to draw undefined optional graphics
Before, the program crashed (segmentation fault) if optional graphics
(like "menu.button_level_number", which are defined to be off-screen
by default) are defined to a valid screen position (by setting the
corresponding screen position values like "main.button.level_number.x/y"
to positions on the screen), but no image was defined to be used for
that graphic.

This bug was fixed by using fallback graphics when attempting to draw
those undefined optional graphics.