fixed selecting correct player in single player mode (after network gaming)
When playing levels that contain more than one player in single player
mode, selecting the actual player when starting the level worked
correctly before only if the game was always played in single player
mode since it was started. It did not work correctly in the following
special case: There was a network multi-player game played before, but
the other (remote) client has disconnected (quit the game), leaving the
other client with a local player that might not be the first player.

In this case, using the first active player (which could be the first
player in the level) for the local player (which could be the second
player, because the network client was started after the now
disconnected client) would result in a broken player mapping, so the
player in the game cannot be moved.

This bugfix solves the problem by not using the first active player,
but the local player.