added setup option to directly play the next level after solving a level
Before, the program returned to the main menu after the player solved
a level (and the high score table was displayed). From there, the next
level could be played by starting the game with the incremented level.

Now, the program automatically starts playing the next level after the
previous level was successfully solved.

This new setup option is enabled by default now, changing the previous
default behaviour of the program.