fixed using private level set if network game level set not found
Before, if a network client started playing a level from a level set
that is not installed for some other network client, that client did
a fallback to playing the default private level set. This caused the
network clients effectively playing completely different levels with
the input actions of the other clients, which does not make any sense.

Now, the network game is stopped for all clients in such a case.