fixed solving level when playing multi-player games
Before, solving a level was broken in both local and network
multi-player games:

In local multi-player games, the level was only solved if the player
marked as "local player" (which is only one of all players in a level)
entered the exit (but not if any other player entered the exit). As an
additional bug, the level was *always* solved as soon as the "local
player" entered the exit, regardless if any other player was still on
the playfield.

In network multi-player games, the level was always solved for the
first player entering the exit (as this player is always the "local
player" in one of the participating network clients), while the game
was frozen on all other network clients (and had to be manually ended
by using the Escape key or stop button).

With this fix, multi-player games will always be solved (locally or on
all network clients) if the last player on the playfield has entered
the exit.