fixed exiting players with custom elements when playing multi-player games
Before, exiting a player using custom element actions only worked in
single-player mode. (In multi-player mode, the level was either
immediately solved if the exited player was the "local player"
(ignoring other players still existing on the playfield), or the
exited player was not removed from the playfield, if it was not the
"local player".)

Now it works correctly also in multi-player games.