fixed extremely nasty bug when using level template in level editor
When checking the checkbox "use template for custom elements" on the
"editor" tab on the level properties page, the playfield area was
replaced with a "broken" level, caused by horrible bugs in the level
initialization function "setLevelInfoToDefaults_MM()". Fixed now.

This bug was introduced with the MM game engine in version 4.1.0.0.