removed storing two strings in function for level filenames
Before, the function "getLevelFilenameFromBasename()" stored two
filenames in static string pointers, one for the level filename, and one
for the corresponding level template filename, because the results were
only used as references before. As they are now stored as string copies
outside if this function, it is sufficient to only store the last result.

This change is directly related to commit 759e0bea.