fixed bug with overwriting level file info from level template
This bug (which was causing crashes due to double free()ing strings)
was a nasty side effect of changing string pointers in the file info
structure from using string references (allocated and freed somewhere
else) to copies of strings allocated and freed for each change, while
still copying the file info from the level template over to the file
info of the level file in function "ActivateLevelTemplate()" (causing
duplicates of string pointers which were therefore freed twice).

This bug was most probably the real cause for the problems fixed by
the changes in commit 759e0bea.