fixed bug with displaying the wrong level number in the game panel
When introducing a global levelset identifier and level number (see
commit b7fe72dc), the level number to be displayed on the game panel
was updated in the main menu and for network games, but not when
selecting another level from the main menu or when automatically
playing the next level after solving a level (and skipping the main
menu before starting the next level), which results in displaying the
wrong level number in the game panel in many cases.

This was fixed now by updating the level number when loading the level
from the level file (moving the function to update the level number to
be displayed from the main menu to the function to load the level).