fixed bug with expiring loop sounds for global animations when playing
When playing the game, loop sounds are automatically expired if they
are not "refreshed" at regular intervals (to prevent playing sounds
for game elements that do not exist on the playfield anymore).

However, this also affects loop sounds for global animations played
during the game, so special care has to be taken to prevent them from
accidentally being expired, too.