added checkbox to player settings in editor if first player solves level
A recently made change (commits 2bf392f1 and 6999c82d) added checking
if all players in a level entered an exit before the level is treated
as being successfully solved, changing the previous behaviour where a
level was solved by the first player entering an exit, which differs
from the behaviour in the EM/EMC game engine, where team-mode levels
require all players to enter an exit before the level is solved.

However, this change broke at least one existing level that was based
on the previous behaviour (level 032 of level set "rnd_falk_sobe"), so
a checkbox in the player settings of the level editor was added to be
able to select the desired behaviour, using the "new" behaviour (all
players must enter an exit) as the default setting.