fixed list of "indestructible" elements as part of "solid" elements
The list of "solid" elements contains a section with "indestructible"
elements that claims to be identical (with one exception) with the
list that only contains the "indestructible" elements, but quite some
elements were missing (that were probably added to the other list at
some time without also being added to the list of "solid" elements).

This change adds these missing elements. (This should be solved in a
more cleaner way, without listing (nearly) the same elements in two
different lists.)

As far as this was tested agains existing solution tapes, this change
does not seem to break any existing levels or tapes.