fixed bug with dragon fire destroying indestructible steel letter walls
Before this change, steel letter walls could be destroyed by dragon
fire, although they are set to be indestructible. This change makes
sure that indestructible elements cannot be destroyed by dragon fire
(with the exception of the "acid" element, which is to be discussed,
but would break level 014 of "rnd_paul_e_collins_iii" if changed).

As far as this was tested agains existing solution tapes, this change
does not seem to break any existing levels or tapes.