added support for (normal and adaptive) vertical sync (vsync)
An option was added to the setup menu to enable vsync for redrawing
the screen in synchronization with the vertical retrace, resulting
in ultra-smooth scrolling on systems where it is supported. To make
use of it in R'n'D together with common LCD displays with 60 Hz screen
refresh frequency, a faster game speed than "normal" has to be chosen,
because "normal speed" in R'n'D means 50 frames per second (which is
the original speed of the game on classic CRT displays, which used a
screen refresh frequency of 50 Hz in Europe by default, and which was
also the original game speed of classic Emerald Mine, for example).

If adaptive vsync is chosen, but not supported, a fallback to normal
vsync is done. (See the SDL wiki for SDL_GL_SetSwapInterval() for a
few more details.)