fixed buggy behaviour when only one player needed to solve a level
Before, when setting "[x] only one player must enter exit", a level
with two players was solved when one player was killed, and then the
other player entered an exit. But it was not yet solved if one player
entered an exit and the other player was still on the playfield.

Now, the first player really solves the level when entering an exit,
regardless of any other players still around on the playfield.