fixed bug with handling networking packets if request dialog is active
Before, it was possible for a network client to start a new game while
another client was still about answering a request dialog (like asking
if a tape for a solved level should be saved), causing synchronization
problems with network packet handling.

With this bugfix, handling network packets is paused during a request.