added support for level sketch copy/paste using Ctrl-c/v in level editor
This enhances the existing support for dumping brushes from the level
editor as level sketches (to STDOUT and/or "stdout.txt", which can
then be added to posts in the R'n'D web forum), as follows:

- Ctrl-c copies the current level (or brush) into the clipboard
- Ctrl-x copies the current level, but as small sized level sketch
- Ctrl-v pastes a level sketch from the clipboard into the current
  level (or brush, if active)

This makes it easy to copy level sketches from the level editor to the
R'n'D web forum, and also from the forum to the level editor (just
temporarily go to the "reply" form for a post containing a level
sketch, then copy the "source code" text of the level sketch to the
clipboard using Ctrl-c, then paste it into the level editor using
Ctrl-v).

This will also make it easy to copy a whole playfield from one level
file to another in the level editor (playfield size will automatically
be adjusted to the size of the pasted playfield).