added support for dynamically sized (and placed) editor palette
When using the value "-1" for columns or rows of the editor palette
(that is, values "editor.palette.cols" and "editor.palette.rows"),
these numbers are dynamically calculated according to the size of the
editor playfield area.

When setting "editor.palette.x" or "editor.palette.y" to "-1", also
the position of the editor palette is dynamically calculated (to be
centered within the playfield area).

(As these values for columns and rows are used at a very early stage
in the level editor to allocate memory for palette buttons, make sure
that no negative numbers are used here for memory calculations.)