fixed bug with CE actions "show envelope" and "exit player" when moving
When using the CE actions "show envelope" or "exit player" with the
condition "after explosion of player", those actions are not performed
if the player was moving at the time of explosion, as they are both
waiting for the player to reach the next tile (which will never happen
in this case).

The best solution is to only check for a "non-moving" tile position of
the player if the player is still active, which will fix the problem.

However, as the check for a "non-moving" tile position is used very
frequently, setting the tile position to "non-moving" for inactive
players was added.