fixed bug with not saving editor cascade settings for two element groups
This bug was caused by adding two new element groups to the level
editor palette area (for Mirror Magic and Deflektor elements), but not
adding them to the list of cascade settings to be saved.

The real problem was that two lists were kept in "src/files.c" for the
settings, and only one was updated when adding the new cascade groups.

To prevent similar bugs in the future, one of the two lists was
removed, as those definitions were not used anyway in this case.