fixed main menu buttons for first level, last level and level number
These three buttons were handled as "main menu screen buttons"
(like the buttons and text labels for the main menu functions),
but should instead be handled as "real gadgets" (just like the
buttons to select previous or next level, but instead triggering
when mouse button is released).

This change was needed to show the button graphics for the active
state of these three buttons when mouse button was held clicked on
them.