fixed graphical bug with drawing tape controls over closed door
This bug happened in two cases where the game was started with the
tape door closed: When starting the game from the "hall of fame"
screen, and when starting the game from the level editor. In these
cases, tape controls were drawn over the closed tape door right
before the door was opened.