fixed level editor copy/paste using Ctrl-c/v for custom/group/DF elements
Before, all custom/group elements were mapped to first custom/group
element when copying them to the clipboard (as this was fine for level
sketches, which do not use any custom artwork anyway), but of course
this breaks copying level playfields in the level editor.

This bug was fixed (which will require updating/extending the level
sketch images for the R'n'D web forum).

Besides this, level sketches from level playfields that contain game
elements with a higher element number than 999 are changed to use a
format of always four digits for all game elements (which will also
require updating/extending the level sketch images for the forum).

Last but not least, while level sketches with four-digit element
number were already created, they were not parsed correctly when
reading them back from the clipboard, which was fixed, too.