fixed handling global animation key actions when mouse button is pressed
When handling key actions for global animations (like pressing "Enter"
or "Escape") which are translated into global animation click events,
they were ignored if a "real" mouse button is already held pressed at
the same time. This can especially happen with animation event actions
that simulate key press events when clicking on global animations, so
the key actions were ignored if the mouse button was held down for too
long.

This change forces key actions of global animations to be handled even
if a mouse button was pressed at the same time.