added new screen fading option ".auto_delay_unit" (currently not used)
This new option will be used to be able to configure the existing
option ".auto_delay" (the time delay after which a title screen will
be automatically faded to the next screen) to use video frames instead
of milliseconds. This can be useful to synchronize global animations
on title screens with the automatic screen fade delay (which may be
inaccurate if using milliseconds for fading, while global animations
always use video frames for configuring animation delay).