fixed handling global animation click events (finally)
Also see commit 4be294f9, which did not fix the problem in all cases,
because the next part of a multi-part global animation (activated by
a click event) will not be active before the following game frame, so
a directly following release event would still just be ignored.

Stopping processing of all following events for one game frame after
handling a click event lets the animation successfully change to the
next part, which can then be triggered by the release event.

An example where this might happen is mentioned in the above commit.