fixed crash bug caused by division by zero
This bug will crash the program when loading a graphics set which uses
toon animation step delay values of zero (instead of positive integers
as it should be).

This bug was introduced around version 4.0.0.0 when adding global
animations (treating toon animations as special global animations),
which could result in animation delay values of zero even though
this was checked and corrected when reading the graphics config file.