fixed redraw bug if gadget buttons are defined outside viewports
This bug was introduced with a fix in commit 50a54138. The problem
with this fix was that it clears REDRAW_ALL from the redraw mask,
which may lead to no redraw at all for this game frame.

This bugfix makes sure that at least the playfield and doors area
will always be redrawn in this case.