added support for gadget buttons defined outside viewports
Commit 50a54138 (and the bugfix in the previous commit) prevent
drawing gadget buttons that are defined outside of viewports (like
playfield or door area), but are placed in the screen border area.

This was not supported in older versions of the redraw system (where
full-screen redraws where always done from a different back buffer).

Recent (SDL2 based) versions of the redraw system do support this, so
this limitation can be removed now to allow for gadget buttons placed
in the screen border area.