fixed bug in editor with GEs using graphic of other element
This bug appeared when using "use graphic of element" and then
changing some other GE settings (like counters etc.), which caused
the custom graphics to be reset to the GEs default graphics.

This is the same bug as the one fixed in commit 77c74cbf, but for
group elements instead of custom elements.